LayoutTests/plugins/npruntime/delete-plugin-within-hasProperty-return-true-expected.txt

CONSOLE MESSAGE: Uncaught ReferenceError: NPObject deleted
